# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/product-api/product/search": {
            "get": {
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品相关"
                ],
                "summary": "门店内商品搜索",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "渠道id",
                        "name": "channel_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "门店编码",
                        "name": "finance_code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "keywords",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页索引",
                        "name": "page_index",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页大小",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.BaseResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}